refactored operations
